% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Inverse logit transformation}
\usage{
expit(z)
}
\arguments{
\item{z}{Numeric vector.}
}
\value{
Numeric vector with values in \eqn{(0,1)}.
}
\description{
Computes the inverse logit of a numeric input.
}
\examples{
expit(c(-1, 0, 1))

}
