% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_srs_val.R
\name{sim_srs_val}
\alias{sim_srs_val}
\title{Create Simple Random Sample (SRS)}
\usage{
sim_srs_val(dat, n)
}
\arguments{
\item{dat}{The complete data to generate SRS samples from}

\item{n}{Sample size}
}
\value{
Dataset with validation indicator `V_SRS` and partially
observed outcome `Y_SRS`.
}
\description{
Create SRS sample from a complete dataset.
}
\examples{
num_features <- 30
pX <- rexp(n = num_features, rate = 6) ## pX follows exponential
betas <- c(-1.5, rep(x = c(-0.5, -0.75, 0.25), times = 7)[-21], rep(x = 1, times = 10))
data <- sim_data(N = 10000, betas = betas, beta_surr = 3, pX = pX, pYstar = 0.1) |> data.frame()

data_srs <- data |> sim_srs_val(n = 500)
head(data_srs)

}
