% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas_gridsearch.R
\name{betas_gridsearch}
\alias{betas_gridsearch}
\title{Grid search for parameters given target TNR and FPR (or TNR)}
\usage{
betas_gridsearch(N, grid, targets, betas_fixed, pX, prop_Y = 0.5)
}
\arguments{
\item{N}{Data size}

\item{grid}{A data frame of candidate parameter combinations. Must
contain the columns `beta0`, `beta_surr`, and `pYstar`.}

\item{targets}{A named numeric vector of target values for
`prevalence`, `TPR`, and `FPR`.}

\item{betas_fixed}{Fixed coefficients in the logistic regression to model Y}

\item{pX}{Marginal feature frequencies (should be in decreasing order)}

\item{prop_Y}{Desired prevalance of the sample, used to calculate Ropt from each combination of betas}
}
\value{
A data frame containing the candidate parameter values, the
corresponding simulated performance measures, and loss values
comparing each candidate to the target metrics. Smaller loss values
indicate closer agreement with the targets.
}
\description{
Performs a grid search over candidate values of the intercept,
surrogate coefficient, and surrogate prevalence to identify settings
that produce simulated data with target outcome prevalence, true
positive rate (TPR), and false positive rate (FPR).
}
\examples{
grid <- expand.grid(
  beta0      = seq(-4, 0, by = 1),
  beta_surr  = seq(-6, 6, by = 1),
  pYstar = seq(0.1, 0.95, by = 1)
)
targets <- c(prevalence = 0.1, TPR = 0.4, FPR = 0.05)
betas_fixed <- c(rep(x = c(-0.5, -0.75, 0.25), times = 7)[-21], rep(x = 1, times = 10))
pX    <- rexp(n = 30, rate = 6)
pX <- sort(pX, decreasing = TRUE) ## sort from least- to most-frequent features
results <- betas_gridsearch(N = 1000, grid, targets, betas_fixed, pX)
head(results, 10)
}
