% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Ropt.R
\name{calc_Ropt}
\alias{calc_Ropt}
\title{Create surrogate-guided sample (SGS)}
\usage{
calc_Ropt(dat, n, npv, ppv, prop_Y = 0.5)
}
\arguments{
\item{dat}{Complete dataset containing `Y` and `Ystar`.}

\item{n}{Validation sample size.}

\item{npv}{Negative predictive value of the surrogate.}

\item{ppv}{Positive predictive value of the surrogate.}

\item{prop_Y}{Target or assumed outcome prevalence. Default is `0.5`.}
}
\value{
Vector of two scalars optimal proportion to validate from surrogate cases (with and without truncation)
}
\description{
Create a surrogate-guided validation sample from a complete dataset
using the surrogate variable `Ystar`.
}
